refactor: move about to modal
