+++ src/components/Modal/Home/SchoolGuide.tsx
