+++ src/components/Modal/index.tsx
@@ -1,3 +1,4 @@
 export { default as BOSSTimetable } from './BOSSTimetable';
-export { default as SchoolGuide } from './Guides/SchoolGuide';
-export { default as WelcomeGuide } from './Guides/WelcomeGuide';
+export { default as About } from './Home/About';
+export { default as SchoolGuide } from './Home/SchoolGuide';
+export { default as WelcomeGuide } from './Home/WelcomeGuide';
